feat: Add ActionDispatch#request_id to rack span as 'action_dispatch.request_id'
ActionDispatch::RequestId is default middleware for Rails, it accepts a
request ID from a header if it exists, otherwise it generates a new
request ID with SecureRandom.uuid.

The default header used is X-Request-Id, though it is possible to
configure which header to look in when initializing the middleware.

ActionDispatch::Request#uuid is an alias for ActionDispatch::Request#request_id

Ref: https://api.rubyonrails.org/classes/ActionDispatch/RequestId.html

As RequestId is the name of the middleware and the default header, this
adds the new attribute 'action_dispatch.request_id' to the rack span.